changed: made zombie pigman of type zombie pigman, not a different kind of zombie added: 2 new tiers of zombie pigman fixed: attack animation for blocks (and nexus) fixed: pigman hammer render full3D() fixed: bug when playing sound to bound players once they left the game

+++ src/main/java/invmod/common/entity/EntityIMThrower.java
@@ -96,7 +96,7 @@ public class EntityIMThrower extends EntityIMMob
 @Override
 	public void onUpdate(){
 		super.onUpdate();
-		if ((this.worldObj.isRemote) && (this.metaChanged != getDataWatcher().getWatchableObjectByte(29))) {
+		if ((this.worldObj.isRemote) && (this.metaChanged !=Byte.valueOf((byte)getDataWatcher().getWatchableObjectByte(29)))) {
 			DataWatcher data = getDataWatcher();
 			this.metaChanged = data.getWatchableObjectByte(29);
 			setTexture(data.getWatchableObjectInt(31));
